♻️ Make createInlineDefenses a Pf2eJsonNodeReader default method
